mesa: Cleanup packaging.
Make some changes to the order of the PACKAGE variable to ensure
that the last package in the list is still "mesa".  This avoids
issues with the debian.bbclass renaming feature changing the package
names to libxatracker-blah.

Also, put all the debug bits into the main mesa-dbg recipe rather
than creating a separate one.

Signed-off-by: Drew Moseley <drew_moseley@mentor.com>