grub: add cmdpath to grub configuration file
An issue was found where when yocto is installed to a hard disk
from a live USB key, no matter what boot order was selected in
the BIOS menu, the grub menu of the USB key would always show
up. The issue was narrowed down to the way grub configuration
file gets embedded into the standalone bootx64.efi image.

The commit prepends ($cmdpath) to the search path for grub.cfg,
thus making sure the UEFI environment itself sets this to the
correct device path. This in-turn lets the grub.cfg of the boot
device to be loaded.

Signed-off-by: Arindam Nath <arindam.nath@amd.com>
Reviewed-by: Drew Moseley <drew_moseley@mentor.com>
Tested-by: Ravikiran Polepalli <Ravikiran_Polepalli@mentor.com>