psplash: fix dependency chaining for psplash-quit
The systemd unit for psplash-quit does not account for any display
manager to be running and this can cause various failures
when the service starts. One major case that was identified
happens in case of Sato GUI.
The xserver-nodm service which is a display manager service
can start before psplash-quit is run and takes over the
framebuffer then psplash-quit starts and tries to clear up the
buffer and fails, leading to a crash to the Sato GUI as well.
We fix this by making sure that psplash-quit is run before
any display manager service is started.

Signed-off-by: Awais Belal <awais_belal@mentor.com>