steppeeagle.conf: provide display mode for VESA
A display mode is necessary in order to kick off the VESA driver
through kernel commandline particularly in the legacy boot mode
otherwise the frame buffer device (/dev/fb) is not created and
hence the logo/splash displays fail.
This is safe as X server can further set itself up according to
its own configuration rather than relying on this.

Signed-off-by: Awais Belal <awais_belal@mentor.com>