mesa: update packaging for libXvMC gallium extensions
The libXvMC gallium extensions of mesa provide a shared lib
libXvMCGallium.so. If the contributing sub-extensions are
packaged separately it affects the rpm packaging in while
generating the package.spec file and generates %post for
the shared lib as many number of times as the sub-extensions
were split. This causes the do_package_write_rpm step
to fail unexpectedly and cause un-seen issues.
We now package these sub-extensions together to avoid
such scenario.

Signed-off-by: Awais Belal <awais_belal@mentor.com>