packagegroup-core-x11: use xserver-common as runtime common
A dependency chain exists between poky and meta-oe which
is not handled properly by the debian packaging system.
This causes the do_rootfs step to fail when debian
packaging is used. The chain is still being reworked
upstream.
To resolve the issue we explicitly define xserver-common
as our VIRTUAL_RUNTIME as this takes care of the funky
dependency chain and copes up with the packaging issue.
Note that xserver-common is what we end up with in other
packaging systems.

Signed-off-by: Awais Belal <awais_belal@mentor.com>