amd-spi: unregister spirom device on exit
The driver registers spirom device on init by using the
spi_new_device call which acquires the chip select pin
through board_info. If the driver does not unregister
this device while exiting the chip select pin never frees
and consequently a second modprobe for the device will
fail with chip select already in use error.
We now keep the device pointer for the device added through
spi_new_device and make an unregister call on the same
on driver's exit.

Signed-off-by: Awais Belal <awais_belal@mentor.com>