linux-yocto: fix jmp behavior for early idt handlers
The kernel hangs while trying to do a load_idt in the
early start up phase and there's nothing obvious to observe
on any of the consoles.
This happens because jmp targets for early_idt_handlers are
global which should always generate a long jmp which in
theory should work correctly but tends to fail with newer
assemblers.

Patch backported from
https://patchwork.kernel.org/patch/6469791/

Signed-off-by: Awais Belal <awais_belal@mentor.com>