linux-yocto: fix CRC generation issue
Due to a faulty CRC generation issue the Module.symvers file
is updated incorrectly for some exported symbols which leads
to build time warnings and runtime failures for loadable
modules.
We now use an upstream kernel patch which fixes the problem
by correctly handling typeof() for generation of kernel
symbols.

Signed-off-by: Awais Belal <awais_belal@mentor.com>