linux-yocto/rt: split out standard only configurations
Some of the configuration being passed through user config fragment
are explicitly disabled in the kernel for RT type. This produces
build warnings that the configuration does not apply cleanly.
We split out the mentioned configuration to a separate config
fragment and use that in case of standard type kernel only.

Signed-off-by: Awais Belal <awais_belal@mentor.com>