libomxil: correctly fix the rpath issue
Earlier we provided the library paths through the -L switch
in our LDFLAGS which caused the compiler automatically embed
wrong (host) rpaths in the test binaries which generated
warnings at build time and other issues on run time.
We fix this properly by utilizing the libtool description
file for the required library produced by the same package
to allow correction in these paths.

Signed-off-by: Awais Belal <awais_belal@mentor.com>