linux-yocto-rt: disable I2S audio
Some components that are required for the I2S audio
to work correctly are part of the AMDGPU driver like
the amd_gnb_bus. The AMDGPU driver is not RT safe yet
so it is disabled which makes the RT builds fail.
We now disable these failing components on the RT
kernel builds to cover up the issue.

Signed-off-by: Awais Belal <awais_belal@mentor.com>