amd-common-configurations.inc: fix resolution issues
The Mentor Embedded splash screen has a higher resolution
than what we get with vga=802 (800x600) so it comes out
clipped. Also the vga=xxx parameter usage has been
deprecated under grub2 and cannot be used to set a
particular resolution so it picks up a default in most
cases.
We now use vga=803 (1024x768) only for handling the
syslinux environment and set gfxpayload accordingly
to cope for the grub2 scenarios.

Signed-off-by: Awais Belal <awais_belal@mentor.com>