amd-common-configurations.inc: fix toaster build artifacts
The Toaster build artifacts collection mechanism depends
on the IMAGE_FSTYPES variable to identify if required
build artifacts are present. The Toaster GUI only shows
the build artifacts if they are detected properly.
We now add the required image types explicitly for
AMD BSPs so Toaster shows the build artifacts as
intended.

Signed-off-by: Awais Belal <awais_belal@mentor.com>