codexl: Resolve race conditions with mkdir.
There is an apparent race condition with the mechanism used
for making directories.  There are corner cases where the
compile step is invoked before the output objdir is created.

This forces all compile steps to be prepended with a mkdir
command.  This may be more than is needed but does seem to
resolve the current races.

Signed-off-by: Drew Moseley <drew_moseley@mentor.com>