v4l-utils: add dependency on libgl and libglu when opengl is enabled
The utility v4l2gl requires gl.h and libgl + libglu libraries
for building cleanly. We need to add an explicit dependency
here to avoid any sort of races although the configure
script checks for availability of the libraries by default
but this can race on the header (gl.h) availability on the
sysroot causing:

| ../../../v4l-utils-1.10.1/contrib/test/v4l2gl.c:27:19: fatal error: GL/gl.h: No such file or directory
|  #include <GL/gl.h>
|                    ^
| compilation terminated.

during do_compile.

Signed-off-by: Awais Belal <awais_belal@mentor.com>