steppeeagle: disable DEBUG_PREEMPT and DEBUG_RODATA
These configurations conflict with KGDB debugging in
some cases and generate stack dumps in various other
situations.
The stack dumps related to mwaitx are properly fixed
through
https://patchwork.kernel.org/patch/8549611/
but the mentioned patch is only available for Yocto
kernel 4.8 and we don't use these so it is much
better to avoid the problem with this for the time
being rather than backporting the patch.

Signed-off-by: Awais Belal <awais_belal@mentor.com>