llvm: fix flags reporting through host llvm-config
Reporting the sysroot from any of the compiler/linker
flags through llvm-config would result in breakage of
the package depending on LLVM as its sysroot would point
to LLVM's sysroot at the end of the day.
We strip out the --sysroot enteries from the
BuildVariables.inc to avoid the above mentioned scenario
and additionally adjust our CXX and LDFLAGS to avoid
duplicating the sysroot variable across the build.

Signed-off-by: Awais Belal <awais_belal@mentor.com>