snowyowl/linux-yocto: fix build when full RT configs are enabled
The RT kernel patches modify the interrupt handling mechanisms
in a way that using do_softirq directly is prohibited. Rather
the support forces the user to use a thread based version for
the same sort of funtionality by calling thread_do_softirq.
This fixes the build issue by calling the appropriate function
depending on the kernel configuration.

Signed-off-by: Awais Belal <awais_belal@mentor.com>