matchbox-keyboard: fix warning for scriptlet failure
The GTK module scriptlet fails to run as it utilizes qemu
which is currently not working for our BSPs due to lack of
a few instructions emulation on the qemu platform and warnings
are generated although the scriptlet is deffered till first
boot. The following warning is seen:

WARNING: Intentionally failing postinstall scriptlets of ['matchbox-keyboard-im', 'eudev-hwdb'] to defer them to first boot is deprecated. Please place them into pkg_postinst_ontarget_${PN} ().

We now defer the scriptlet forcefully till the first boot to
cover up for the warning.

Signed-off-by: Awais Belal <awais_belal@mentor.com>