tune-x.inc: use the m64 TUNE_FEATURE
All the currently supported BSPs are 64bit so use the
appropriate m64 TUNE_FEATURE in order to get the build
environment evaluated properly. Otherwise the basic -m64
switch which is required for proper linker output
specification and other such things do not end up in
the correct form and the build fails in case a toolchain
that supports 32 and 64bit builds is used.

Signed-off-by: Awais Belal <awais_belal@mentor.com>