boost: fix fuzz in the patch to avoid QA warning
The new bitbake scheme does not allow fuzzes in the
patches and throws a warning if one is seen. This fixes

WARNING: boost-1.59.0-r0 do_patch:
Some of the context lines in patches were ignored. This can lead to incorrectly applied patches.
The context lines in the patches can be updated with devtool:
...
Applying patch boost-math-disable-pch-for-gcc.patch
patching file libs/math/build/Jamfile.v2
Hunk #1 succeeded at 13 with fuzz 1.
...

Signed-off-by: Awais Belal <awais_belal@mentor.com>